Request log summary model
- Added request log summary model
